fix(agent-sdk): run post-creation setup in EnterWorktree tool
enterWorktreeTool created worktrees without copying .wave/settings.local.json
and .worktreeinclude-listed files (.env/.mcp.json), unlike the CLI -w path.

Move performPostCreationSetup from packages/code into agent-sdk worktreeUtils
(async, single source of truth) and call it best-effort after creating a new
worktree. packages/code now reuses the SDK implementation.
