+++ packages/code/src/utils/worktree.ts
@@ -2,7 +2,11 @@ import { execFile } from "node:child_process";
 import { promisify } from "node:util";
 import * as path from "node:path";
 import * as fs from "node:fs";
-import { getDefaultRemoteBranch, getGitMainRepoRoot } from "wave-agent-sdk";
+import {
+  getDefaultRemoteBranch,
+  getGitMainRepoRoot,
+  performPostCreationSetup,
+} from "wave-agent-sdk";
 
 // Never use execFileSync here: the shared `wave --stdio` process handles all
 // desktop sessions, so a synchronous git call (especially a multi-second
@@ -49,338 +53,6 @@ export function validateWorktreeSlug(name: string): void {
       );
     }
   }
-}
-
-// --- .worktreeinclude matching ----------------------------------------------
-
-interface WorktreeIncludePattern {
-  /** Normalized pattern text (no leading "/", no trailing "/", no "!") */
-  raw: string;
-  negated: boolean;
-  /** Pattern ends with "/" — matches directories only */
-  dirOnly: boolean;
-  /** Pattern had a leading "/" — anchored to the repo root */
-  anchored: boolean;
-  /** No "/" and not anchored — matches the basename at any level */
-  anyLevel: boolean;
-  regex: RegExp;
-}
-
-/** Translate a single gitignore glob into a RegExp (gitignore semantics). */
-function globToRegExp(glob: string): RegExp {
-  let source = "^";
-  for (let i = 0; i < glob.length; i++) {
-    const ch = glob[i];
-    if (ch === "*") {
-      if (glob[i + 1] === "*") {
-        if (glob[i + 2] === "/") {
-          // "**/" matches zero or more leading directories
-          source += "(?:.*/)?";
-          i += 2;
-        } else {
-          source += ".*";
-          i += 1;
-        }
-      } else {
-        source += "[^/]*";
-      }
-    } else if (ch === "?") {
-      source += "[^/]";
-    } else if (ch === "[") {
-      const end = glob.indexOf("]", i + 1);
-      if (end === -1) {
-        source += "\\[";
-      } else {
-        let charClass = glob.slice(i, end + 1);
-        if (charClass.startsWith("[!")) {
-          // gitignore uses "[!...]" for a negated character class
-          charClass = "[^" + charClass.slice(2);
-        }
-        source += charClass;
-        i = end;
-      }
-    } else {
-      source += ch.replace(/[.+^${}()|\\]/g, "\\$&");
-    }
-  }
-  source += "$";
-  return new RegExp(source);
-}
-
-/**
- * Parse .worktreeinclude content into patterns. Blank lines and "#" comments
- * are skipped, matching gitignore conventions.
- */
-function parseWorktreeIncludePatterns(
-  content: string,
-): WorktreeIncludePattern[] {
-  const patterns: WorktreeIncludePattern[] = [];
-  for (const rawLine of content.split(/\r?\n/)) {
-    const line = rawLine.trim();
-    if (!line || line.startsWith("#")) continue;
-    let raw = line;
-    const negated = raw.startsWith("!");
-    if (negated) raw = raw.slice(1);
-    const dirOnly = raw.endsWith("/");
-    if (dirOnly) raw = raw.slice(0, -1);
-    const anchored = raw.startsWith("/");
-    if (anchored) raw = raw.slice(1);
-    if (!raw) continue;
-    patterns.push({
-      raw,
-      negated,
-      dirOnly,
-      anchored,
-      anyLevel: !anchored && !raw.includes("/"),
-      regex: globToRegExp(raw),
-    });
-  }
-  return patterns;
-}
-
-/** Test a single pattern against one candidate path. */
-function testPatternAgainst(
-  pattern: WorktreeIncludePattern,
-  candidate: string,
-): boolean {
-  if (pattern.anyLevel) {
-    // Match the full path or any "/"-suffix (basename at any level)
-    if (pattern.regex.test(candidate)) return true;
-    for (let i = 0; i < candidate.length; i++) {
-      if (candidate[i] === "/" && pattern.regex.test(candidate.slice(i + 1))) {
-        return true;
-      }
-    }
-    return false;
-  }
-  return pattern.regex.test(candidate);
-}
-
-/**
- * Test a pattern against a path and every ancestor prefix. Excluding a
- * directory excludes everything beneath it (gitignore semantics), so ancestor
- * prefixes are always treated as directories.
- */
-function patternMatchesPath(
-  pattern: WorktreeIncludePattern,
-  relPath: string,
-  isDir: boolean,
-): boolean {
-  let candidate = relPath;
-  for (;;) {
-    const candidateIsDir = candidate === relPath ? isDir : true;
-    if (
-      !(pattern.dirOnly && !candidateIsDir) &&
-      testPatternAgainst(pattern, candidate)
-    ) {
-      return true;
-    }
-    const idx = candidate.lastIndexOf("/");
-    if (idx === -1) break;
-    candidate = candidate.slice(0, idx);
-  }
-  return false;
-}
-
-/** Last-match-wins resolution with "!" negation. */
-function worktreeIncludeMatches(
-  patterns: WorktreeIncludePattern[],
-  relPath: string,
-  isDir: boolean,
-): boolean {
-  let matched = false;
-  for (const pattern of patterns) {
-    if (patternMatchesPath(pattern, relPath, isDir)) {
-      matched = !pattern.negated;
-    }
-  }
-  return matched;
-}
-
-/**
- * A positive pattern targets something inside a collapsed directory: either it
- * literally starts with the directory path, or the directory path starts with
- * the pattern's literal (pre-glob) prefix.
- */
-function needsExpansion(
-  patterns: WorktreeIncludePattern[],
-  dir: string,
-): boolean {
-  return patterns.some((p) => {
-    if (p.negated) return false;
-    if (p.raw.startsWith(dir + "/")) return true;
-    const globIdx = p.raw.search(/[*?[]/);
-    if (globIdx > 0 && dir.startsWith(p.raw.slice(0, globIdx))) return true;
-    return false;
-  });
-}
-
-// --- Post-creation setup -----------------------------------------------------
-
-/**
- * Copy <repoRoot>/.wave/settings.local.json into the worktree so local
- * configuration (permissions, env, ...) carries over. Missing files are
- * skipped silently; other failures only log a warning.
- */
-async function copyLocalSettingsToWorktree(
-  repoRoot: string,
-  worktreePath: string,
-): Promise<void> {
-  const relativePath = path.join(".wave", "settings.local.json");
-  const sourcePath = path.join(repoRoot, relativePath);
-  const destPath = path.join(worktreePath, relativePath);
-  let content: string;
-  try {
-    content = await fs.promises.readFile(sourcePath, "utf8");
-  } catch (error: unknown) {
-    if ((error as NodeJS.ErrnoException).code !== "ENOENT") {
-      console.warn(`Failed to read ${sourcePath}: ${(error as Error).message}`);
-    }
-    return;
-  }
-  try {
-    await fs.promises.mkdir(path.dirname(destPath), { recursive: true });
-    await fs.promises.writeFile(destPath, content);
-  } catch (error: unknown) {
-    console.warn(
-      `Failed to copy ${relativePath} into worktree: ${(error as Error).message}`,
-    );
-  }
-}
-
-/**
- * Copy files listed in <repoRoot>/.worktreeinclude into the worktree. Each
- * non-comment line is a gitignore pattern; matching gitignored files are
- * copied at their relative paths.
- */
-async function copyWorktreeIncludeFiles(
-  repoRoot: string,
-  worktreePath: string,
-): Promise<string[]> {
-  const includePath = path.join(repoRoot, ".worktreeinclude");
-  let content: string;
-  try {
-    content = await fs.promises.readFile(includePath, "utf8");
-  } catch {
-    return [];
-  }
-  const patterns = parseWorktreeIncludePatterns(content);
-  if (patterns.length === 0) return [];
-
-  const worktreeRelPath = path
-    .relative(repoRoot, worktreePath)
-    .split(path.sep)
-    .join("/");
-  // The worktree lives under .wave/worktrees/, which is gitignored — never
-  // copy the worktree (or anything inside it) into itself.
-  const isSelf = (entry: string) =>
-    entry === worktreeRelPath || entry.startsWith(worktreeRelPath + "/");
-
-  // "--directory" collapses fully-ignored directories into single entries,
-  // which keeps the listing fast on large repos.
-  let entries: string[];
-  try {
-    const { stdout } = await execFileAsync(
-      "git",
-      [
-        "ls-files",
-        "--others",
-        "--ignored",
-        "--exclude-standard",
-        "--directory",
-      ],
-      { cwd: repoRoot, encoding: "utf8" },
-    );
-    entries = stdout.trim().split("\n").filter(Boolean);
-  } catch {
-    return [];
-  }
-
-  const copied: string[] = [];
-  const dirsToExpand: string[] = [];
-
-  const copyToWorktree = async (relativePath: string): Promise<void> => {
-    const srcPath = path.join(repoRoot, relativePath);
-    const destPath = path.join(worktreePath, relativePath);
-    try {
-      await fs.promises.mkdir(path.dirname(destPath), { recursive: true });
-      await fs.promises.copyFile(srcPath, destPath);
-      copied.push(relativePath);
-    } catch (error: unknown) {
-      console.warn(
-        `Failed to copy ${relativePath} into worktree: ${(error as Error).message}`,
-      );
-    }
-  };
-
-  for (const entry of entries) {
-    if (isSelf(entry)) continue;
-    if (entry.endsWith("/")) {
-      const dir = entry.slice(0, -1);
-      if (worktreeIncludeMatches(patterns, dir, true)) {
-        // The directory itself matches — copy it wholesale
-        try {
-          await fs.promises.cp(
-            path.join(repoRoot, dir),
-            path.join(worktreePath, dir),
-            {
-              recursive: true,
-            },
-          );
-          copied.push(entry);
-        } catch (error: unknown) {
-          console.warn(
-            `Failed to copy ${entry} into worktree: ${(error as Error).message}`,
-          );
-        }
-      } else if (needsExpansion(patterns, dir)) {
-        dirsToExpand.push(dir);
-      }
-    } else if (worktreeIncludeMatches(patterns, entry, false)) {
-      await copyToWorktree(entry);
-    }
-  }
-
-  if (dirsToExpand.length > 0) {
-    // List the files inside collapsed dirs whose contents are targeted
-    try {
-      const { stdout } = await execFileAsync(
-        "git",
-        [
-          "ls-files",
-          "--others",
-          "--ignored",
-          "--exclude-standard",
-          "--",
-          ...dirsToExpand,
-        ],
-        { cwd: repoRoot, encoding: "utf8" },
-      );
-      for (const file of stdout.trim().split("\n").filter(Boolean)) {
-        if (isSelf(file)) continue;
-        if (worktreeIncludeMatches(patterns, file, false)) {
-          await copyToWorktree(file);
-        }
-      }
-    } catch {
-      // Expansion is best-effort; the worktree is already usable without it
-    }
-  }
-
-  return copied;
-}
-
-/**
- * Set up a freshly created worktree: copy local settings and gitignored
- * project files (via .worktreeinclude) from the main repo. Best-effort — any
- * failure only logs a warning and never fails worktree creation.
- */
-export async function performPostCreationSetup(
-  worktreePath: string,
-  repoRoot: string,
-): Promise<void> {
-  await copyLocalSettingsToWorktree(repoRoot, worktreePath);
-  await copyWorktreeIncludeFiles(repoRoot, worktreePath);
 }
 
 /**
